[DE1-SoC] Simplified Nios demo program & HPS linux demo program & HPS baremetal demo programs.
No longer have hex displays, because their code was extremely complex for the simple demonstration intended to be done here. Simple strobing effect on the LEDS are done instead. The LEDS are divided among the Nios and the HPS, and each has its own portion for their corresponding strobing effects.